Make processValidation method utility
